fix company object initialization

+++ src/main/java/com/hakaneroztekin/glassdoorScraper/Company.java
@@ -8,5 +8,7 @@ import org.springframework.stereotype.Component;
 @Component
 public class Company {
     String title;
-    Integer rate;
+    Double rate;
+
+
 }
